Fixing store AsyncStore isFirstLaunch key and rendering flip-card
